Attempt to allow more time for testMemcachedFailover
This will cap each attempt at a maximum of 10 seconds to account for
CPU-side lag wherein the failover does not take place in a timely
manner. This may still fail at some points, but the failure counts will
hopefully be less. The main intent is to allow slower CI builders to
succeed with this test.

Change-Id: Ieb9117b5d36f85ffc276a9bc552a7058cddd97ca
Reviewed-on: http://review.couchbase.org/47949
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>