CCBC-606: Modify packet's start time to flush time (optionally)
This resets the packet's start time to its flush time if readj_ts_wait
is enabled. This makes the packet's actual timeout start from the point
in time when it was placed on the wire rather than the point in time in
which it was scheduled (or when lcb_wait() was invoked).

Accordingly, the retry queue will no longer inspect the 'start' field,
preferring to copy the start field when the packet is first placed
inside the retry queue. This prevents packets from being infinitely
retried (if they are sent and their start time gets updated).

Change-Id: I2e31442ac49563dbda890046d8411295469c0025
Reviewed-on: http://review.couchbase.org/50464
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>