Remove mcserver_alloc2
While the intent here may have been to decouple the creation of the
server with the configuration, this has led to some weirdness when
examining the code. The fact that we set the new configuration
immediately after receiving it makes the need for this function
redundant.

If this behavior changes in the future, we'll add a new param instead.

Change-Id: Iead5df40b3586c14fc9ad0a247e392e25eef1aaf
Reviewed-on: http://review.couchbase.org/51455
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>