Pillowfight: Adjust key generation so negative keys are not created
In generateKey() the sequence number is adjusted by the KeyGenerator's
offset minus 1; however in the case of a single thread there is just
one KeyGenerator and so offset is zero. This causes the seqno to be
decremented by 1, which when it's already zero gives a negative
key.

Remove the -1, and reorder where op.seqno is updated to ensure that
keys are now generated in the range [0...N-1] - previously they would
be in the range [1...N-1,-1].

Change-Id: I5c391b3ad67b23407e027bc0523556a38193692d
Reviewed-on: http://review.couchbase.org/52978
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>