CCBC-633, CCBC-632: Prepared statements and more integration
This rewrites the N1QL handling system in C++ and makes it JSON-aware.
Statements passed to lcb_n1ql_query are examined and necessary
parameters are injected as necessary.

Prepared statement support is also added by creating an internal "plan
cache" which can be used with a simple command flag. This cache also
features an LRU.

Finally, retry logic is added (no way to test this currently) in case a
statement must be re-prepared again.

Change-Id: I998ba13283ff52492a99b629dc0c02a34e06ff26
Reviewed-on: http://review.couchbase.org/54197
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>