CCBC-681: Use config throttling on bodyless NMVB with CCCP
If CCCP is enabled then a bodyless NMVB may be a sign of a server-side
optimization

The current behavior is to immediately request a new config from the
cluster if the currently-received NMVB does not contain a config.

The new behavior will make the aforementioned request subject to
throttling/rate limiting restrictions (by default, 10 seconds) if the
current configuration was received via CCCP (and thus we are certain
that CCCP is being used).

Change-Id: Id9d519a6bfec070bef092dc4ac09877795310591
Reviewed-on: http://review.couchbase.org/61542
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>