CCBC-525: Fix non-deterministic reconfigureOnNodeFailover test
We ignore the result if the instance cannot properly detect the change
in node counts. The block becomes as isolated as possible to detecing a
configuration change (so we're sure it's not crashing, hanging, or
otherwise).

A message will still be printed to the screen if detection fails,
however it will not result in an overall test failure.

Change-Id: I9056c898d2452a82e3f59c313e6ecb735706be36
Reviewed-on: http://review.couchbase.org/61579
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>