CCBC-638: add cbft row api
This adds the basic row API for cbft. While the actual server-side CBFT
API is subject to change, the API exposed in libcouchbase is very
similar to the existing view and n1ql API.

Currently no helper classes have been defined to create helper
constructors for the query JSON.

This also fixes some bugs in the vbucket implementation of the FTS
service discovery with respect to the actual query URL, which it seems
is actually dependent on the index being queried.

Change-Id: Icaa70883e8fad480c6bc9e7bd339466a68be788d
Reviewed-on: http://review.couchbase.org/62203
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>