SSL: Allow to bypass openssl initialization
By default we set up OpenSSL initialization and locking routines. This
is usually ok, but conflicts with programs which already setup SSL in
their own right (for example, php). While we can't determine if OpenSSL
is actually loaded, we _can_ at least offer an escape hatch so that the
application does not crash.

This functionality can be enabled by using `ssl=no_global_init` in the
connection string. Note that SSL connection string options are
_additive_ as well.

Change-Id: Ie50f2c0cf67333e7ac8be6f9dbd5fe54e6480e42
Reviewed-on: http://review.couchbase.org/64733
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>