CCBC-704: Don't get stuck in loop when single server fails
This ensures that we don't end up in a situation where the single-server
socket is never actually cleaned up. This is fixed by properly handling
entered/exited events and the pending counter

Previously the pending counter was decremented only *after* the Cw_ex
callback, making the application code (in this case, mcserver) never
actually perform the cleanup.

Change-Id: I0717c632adea7d1a6aaf1dcc1516104885ca917b
Reviewed-on: http://review.couchbase.org/64855
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: buildbot <build@couchbase.com>