Use ntohll/htonll for CAS values
This makes it so that CAS is always encoded/decoded properly according
to network byte order. This was previously not an issue as CAS was
essentially an application-local opaque value. This has changed because
N1QL now also exposes the CAS and it is expected to be able to
interoperate between the two.

Change-Id: I63fb9670fc48ff469e0c1dcaaeab23fc638ef418
Reviewed-on: http://review.couchbase.org/66558
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>