CCBC-718: add SUBDOC_F_MKDOCUMENT and GET_COUNT
This adds the GET_COUNT and MKDOC options. `GET_COUNT` retrieves the
number of items in an array or dictionary; MKDOC is another flag (like
MKDIR_P) which will create the document if it does not exist.

This commit also includes tests for the features, alongside an upgrade
to a newer CouchbaseMock version

Change-Id: I1fd5a9a910d37f9933f80b0297864683e9572bbd
Reviewed-on: http://review.couchbase.org/67213
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>