CCBC-758: Select bucket using SELECT_BUCKET if enabled by the server.
This is necessary to support RBAC. In this scenario, the bucket with the
appropriate name is selected for use.

This commit consists of:

* Sending the `SELECT_BUCKET` command if indicated by the HELLO feature
  received.
* New lcb_cntl interfaces (`LCB_CNTL_SELECT_BUCKET`) and connection
  string interfaces to forcefully disabled the sending of the
  `SELECT_BUCKET command (for testing purposes).

Currently the SELECT_BUCKET feature is off by default: Since this is a
released version, we wouldn't want someone using this version in the
future together with a newer version of the server that may potentially
implement SELECT_BUCKET/RBAC differently. Because this feature is
experimental, it must currently therefore be turned on manually.

Change-Id: I3d7d51075baf97c19beedd1bd6cc5bb52f8ee2cd
Reviewed-on: http://review.couchbase.org/74622
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>