CCBC-690: Allow to enable TCP keepalive
This allows the use of the `TCP_KEEPALIVE` socket option to enable
sending keepalive probes on the connection.

Note that it is often insufficient to simply enable TCP Keepalive, and
that kernel-level settings must be adjusted as well - such as the time
to wait before sending keepalive probes (the kernel default - per TCP
standard - is 2 hours, or 7200 seconds).

To make use of this feature, use the `LCB_CNTL_TCP_KEEPALIVE` option in
`lcb_cntl()` or the `tcp_keepalive` option in the connection string.

The value of this option is a boolean.

Change-Id: I665ed4c3224f5a5147ada74501660dcde6874f5e
Reviewed-on: http://review.couchbase.org/77054
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>