Clarify error message when we can't directly map errno to lcb
Previously there used to be a "FIXME" message which gave people cause
for alarm (and in general.. this message is not considered normal) -
however, this is usually not the cause of an error but a symptom
thereof. Additionally, we can't anticipate every and any errno that any
OS may throw at us in any environment, so in many cases there's nothing
to "Fix" in the library.

Change-Id: I4cdf528f92f5608220314d2964382b126f3411d6
Reviewed-on: http://review.couchbase.org/77052
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>