CCBC-732: Fix the request rate throttling
In an attempt to switch from the per-thread limit to the absolute
limit, we've achieved the opposite of what was intended.

For instance, if the "num-threads" option is set to 5 and the
"rate-limit" option is set to 10000, the total throughput will be
250K ops/sec (i.e., a quadratic function of the number of threads).

This patch makes the rate limit absolute by increasing the sleep
time proportionally to the number of threads.

Change-Id: Icbe3080832f9de8f41484a44677ddb9ee5a138f1
Reviewed-on: http://review.couchbase.org/77300
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavel Paulau <pavel.paulau@gmail.com>