Fix inadvertent casting to bool
There are a casting to bool from an int return value. The int was a
tri-state variable however (0/1/-1), but because it was casted to bool
(presumably unsigned int), it got converted to 1/0, and thus -1 became
1.

This caused some tests to fail .. and likely more issues.

This commit also changes the naming from is_closed to check_closed.
Hopefully this will make it a bit clearer that this is not a boolean
function.

Change-Id: Iaaf0a516a732e13ef4f22915d2d7d257849cea7c
Reviewed-on: http://review.couchbase.org/77439
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Build Bot <build@couchbase.com>