CCBC-799: Infer multimode in constructor
Motivation
----------
lcb_subdoc3() perform argument validation based on mode, in
particular it checks that setting expiration is not allowed
in multi-lookup mode. But the mode in MultiBuilder initialized
later, and by the time of the check, it always 0, making it
impossible to specify expiration even in multi-mutation mode.

Modification
------------
In MultiBuilder constructor try to infer mode from first
spec, if it is exist, and fallback to zero otherwise.

Result
------
Multi-mutations can set the expiration.

Change-Id: Ia81f181cdda3905521a781078aded1ab55989100
Reviewed-on: http://review.couchbase.org/81819
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>