CCBC-823: Fix cbc-subdoc/upsert without XATTR
Motivation
----------
With `upsert` subcommand (fulldoc set), it is mandatory to send xattr
mutation, to workardound this limitation, cbc-subdoc generates
`_cbc.version` xattr when the user does not specify `--xattr` switch
with upsert. The problem is that scope of the path string is too narrow
and it does not live long enough to be sent to network, and the library
sending garbage, which results in 0x3d status code.

Modification
------------
Increase scope of the path variable.

Change-Id: I7e2ba91e2e3b7301939811585766a0ef3c8cf00f
Reviewed-on: http://review.couchbase.org/82445
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>