CCBC-831: Fix interface stability of couchbase.h and cntl.h
Motivation
----------
Many of the interfaces in couchbase.h and cntl.h have not been upgraded
to the expected stability level over time. This would be confusing to
a user as important APIs are listed as not committed.

Modifications
-------------
Remove the @uncommitted and @volatile tags on things that should be
committed. Upgrade the histogram interface from @private (which was
incorrect for other reasons) to @volatile. Replace most @private
labled API with @internal, correcting to the actual taxonomy. Update
references in the headers to official documentation.

Result
------
Stability of interface is more clear to internal and external users.

Change-Id: Icc633cfcd9f780e8b8e66033d540e43be8707c27
Reviewed-on: http://review.couchbase.org/82837
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>