CCBC-830: Correct documentation on adjusting time values and timeouts
Motivation
----------
Over the course of time, settings have been added to the library
that are not strictly related to timeouts, but documentation in
the header and doc blocks referred to everything as if it was a
timeout.  Also, documentation did not cover how strings were
handled by the library.

Modifications
-------------
Updated the cntl.h header to indicate that these are generically
time values.  Also updated the static cntl.cc while at it, as it is
only used internally. Updated the anchor and references within the
header as well.

Oh, and in case you want to quibble with the term 'fractional-part'
and suggest mantissa should have been used instead, I humbly submit
this stack exchange discussion:
https://math.stackexchange.com/questions/64042/what-are-the-numbers-before-and-after-the-decimal-point-referred-to-in-mathemati

Result
------
Documentation is clearer about how time values are parsed and used
in libcouchbase settings.

Change-Id: I4fc5b87fea8fa8e702a13dc7c8ce07ad9f9e1cc6
Reviewed-on: http://review.couchbase.org/82842
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>