tools: implement cbc-proxy
Motivation
----------
cbc-pillowfight is a good tool for stress testing, but does not
fit good to simulate applications, which might have bigger idle
times. Having a program, which could run for a long time helps
to find and reproduce issues with connection pooling and topology
changes.

Modification
------------
Implement cbc-proxy command, which uses Forward API of libcouchbase
and exposes interface for dumb memcached clients.

Result
------
It is possible to write simple test scripts without Couchbase SDK.
As a bonus cbc-proxy is an application demonstrating how to use
Forward API of libcouchbase

Change-Id: If2e6cfe3edcfdfaa290b4e93e6f8fb16aab97761
Reviewed-on: http://review.couchbase.org/83444
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>