CCBC-878: Support collections in cbc-pillowfight
Note that this change does not expose anything related to Collections API
for libcouchbase. It defines hidden switches for pillowfight tool to
allow benchmark of collections. The switches are not documented and might
be removed in the future. Use with care.

Generate only "beer:<seqno>" keys:

  cbc pillowfight --separator : --collection beer

Using many --collection will alternate in generating "beer:<seqno>",
"brewery:<seqno>" keys (default separator is ":"):

  cbc pillowfight --collection beer --collection brewery

Change-Id: I11a18df1428eb1d1467a1497456c4d782e0c8007
Reviewed-on: http://review.couchbase.org/86181
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>