Relax assertion in select IO plugin destructor
It is possible that the code will be stopped in the middle of event loop
(and this is possible in PHP for example with max_execution_time limit).
The application might detect this case, and still try to destroy the
instance with IO plugin. It should be possible to cleanup resources for
select plugin. Moreover other plugins allow to do it and stop the IO
loop.

Change-Id: Ie3b4b690e292ff45a003cfac5d859917a38bef5e
Reviewed-on: http://review.couchbase.org/87386
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>