CCBC-925: Do not register nonexistent crypto providers during lookup
The std::map [] operator on the instance->crypto
map was adding a NULL entry to the map when looking up
provider names.

This was leading to a crash on shutdown as the reference counts
were being decremented on NULL entries.

This fix checks for the existence of a provider using std::map::find
rather than using the [] operator.

Change-Id: I86f0b804b85fe4b9182de8f15fbf62f226e893e6
Reviewed-on: http://review.couchbase.org/93754
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>