CCBC-975: Make sure kv service is only enabled if in nodes list.
Motivation
----------
The SDK needs to iterate the nodesExt as part of the config because
some global services (i.e. N1QL) are not to be found in the
traditional "nodes" list of the config. With KV, this turns out
to be a problem during rebalance-in scenarios.

Especially during delta-node recovery and memcached buckets, the
problem amplifies because nodesExt holds all nodes that are part
of the cluster with their services, while the nodes list only
holds the nodes that are currently active for the given bucket.

If the SDK only uses nodesExt (which it does now), it tries to
connect to kv buckets immediately as they become available, even
if they are not active for this bucket yet. For couchbase and
ephemeral buckets this is not a big issue because there the
movement of vbuckets gradually moves over the load, but with
memcache buckets as soon as the node pops up in the config
it is part of the ketama ring and ops are dispatched to it
right away.

Changes
-------
The part of the config loading where we iterate nodesExt now
explicitly checks if a corresponding entry in the nodes list is
available. If not then the node's KV ports will be reset marking
it as if it does not have KV service enabled.

Change-Id: Ia09c7d5cdcdb592d36fb17397412926236ce6afd
Reviewed-on: http://review.couchbase.org/99475
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>