CCBC-988: Update cbc-pillowfight to work with by-id collections
1) Server no longer uses separator, so removed
2) Collections are specifed by leb128 encoded the collection
identified.

The update changes --collection so that it accepts a base16 ID,
the same base which the collections JSON config uses.

Note: This patch does not adhere to the exact client requirements
for collections as we are vbucket hashing the entire key-bytes, the
leb128 prefix and the logical key. Real clients should not do that
so we could in the future allow for a server side move-key operation
which assigns a key to a new collection, but for the purpose of
load-generation this is OK.

Change-Id: I4ead895c3af0c48fa10974ef4b6eb9acabde2943
Reviewed-on: http://review.couchbase.org/100559
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>