CCBC-1092: add count/interval options for cbc-ping
With this fix, cbc-ping will run indefinitely by default with 1 second
interval between requests.

`--interval,-i` switch controls interval in seconds

`--count,-c` sets the number of requests to be sent before exit. In this
mode, the output will also include "[" in the beginning, "]" in the end,
and "," between responses to represent results as well-formed JSON
array.

Change-Id: I81577889b4c2263a5797c84da73f240ad9da0eff
Reviewed-on: http://review.couchbase.org/114309
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>