Make ciphers and minimum TLS version tunable
Allow the user to specify the minimum TLS version to
use and which ciphers to use through the environment variables.

TLS is set through LCB_SSL_MINIMUM_TLS which may be set to:

   tlsv1.1
   tlsv1.2
   tlsv1.3

If set to an invalid/unsupported value the minimum version is
set to TLS1.

Unfortunately OpenSSL use two different API calls in order to
specify the ciphers to allow for the various TLS levels, and
the methods work sligtly different so it has to be tuned
through two different environment variables.

Ciphers for TLS < 1.3 is set through LCB_SSL_CIPHER_LIST and
for TLS >= 1.3 it is set through LCB_SSL_CIPHERSUITES (the
names in use match what OpenSSL calls them).

Two new error codes was added to the system:

LCB_SSL_INVALID_CIPHERSUITES - We failed to set the ciphersuite
LCB_SSL_NO_CIPHERS - OpenSSL was not able to select a single
                     cipher from the list provided.

Change-Id: I48be0d943a360ec1008e17535b791bc2735c2f26
Reviewed-on: http://review.couchbase.org/114677
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>