cbc-n1qlback: Remove use of deprecated function std::random_shuffle
std::random_shuffle is deprecated in C++14, removed in C++17. Replace
with std::shuffle.

Note: When building as part of Couchbase Server with GCC / Clang and
C++17, this function _isn't_ flagged as deprecated as libcouchbase
sets it's own C++ standard (see
cmake/Modules/GetLibcouchbaseFlags.cmake) which overrides the
top-level server setting. As such, this use of std::random_shuffle is
only flagged when building for a non-GCC/Clang compiler, e.g. MSVC.

Change-Id: Ib9b6fcecb8ecff79db9335f0c0a2c77361b18609
Reviewed-on: http://review.couchbase.org/122416
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>