CCBC-1251: No longer log error when using GCCCP on pre-6.5
Motivation
----------
Couchbase Server prior to 6.5 generate a NO_BUCKET error when
attempted to use GET_CLUSTER_CONFIG before a bucket is selected.
This should print a warning rather than an error in the logs.

Changes
-------
Added handling of the NO_BUCKET error when there is intentionally
no bucket selected in CCCP handling.

Change-Id: I050758cba468635774ffbf38acd57c39ed815422
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/129053
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>