Several documentation issues addressed
Motivation
===========
Get api documentation ready to be published.  Maybe not everything,
but at least enough. This covers CCBC-1240, CCBC-1243, CCBC-1241,
CCBC-1245.

Modification
============
Search, Query, and Analytics all were documented in 2.10.6, however that
was lost in the conversion to 3.0.  This just resurrects that documentation,
but adjusts it to match the 3.0-isms.

The KV documentation was all there, but still had 2.x examples for the most
part.  Now, they should be all converted to 3.x.

Also, there were some tag issues that made the docs render poorly, and those
seem to be addressed.

Results
=======
Better api docs.  Now, is it complete?  No.  But better.  Worth looking at and
deciding if it is close enough to publish.  But much more should be done in any
case.

Change-Id: Ic82d7e3f5f1dfd108528ec31356dc6a21da5ab9d
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/130086
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>