CCBC-1278: Fix stripping payload over TLS connection
The library might propagate zero return code from recv() even when
previous calls of recv() successfully transferred bytes to SSL context.
In this case we might lose these trailing bytes and just close the
socket.

The easiest way to reproduce is to call some management API
(cbc admin /pools). In this case the server sends us piece of data and
closes the socket, but libcouchbase receives data in two buffers. First
with headers it handles properly, while the body is getting lost,
because of socket shutdown.

Change-Id: I12187e278c907c0177328a658786079c3d6d4ba7
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/131467
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>