CCBC-1412: Fixed lifetime error for collection/scope in error contexts.
Motivation
----------
A previous change in the handling of collection path handling caused the
scope of the collection_path that is used in the error context to not
outlive the callback invocation, causing the application to receive
possibly corrupted data.

Changes
-------
Moved collection_path scope up a level to ensure it survives long enough
for the callback invocation to complete.

Change-Id: I77d59be0d217d24adc13d372fd6c6ce314cee2ba
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/154689
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>