CCBC-1362: Support for metrics
Add support for metrics.  When enabled, by default will output a
histogram of latencies for various kv operations plus query, search,
analytics and view queries.

If desired, an external metrics collector can be passed in when the
instance is created, which will then be called with latencies for each
operation.  This can be used to call into an opentelemetry meter for
aggregation and export. See /examples/otel_metrics.cc for an example.

Change-Id: Ic702fd3df19d0cba466b83e10c3f8cf66a616404
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/155226
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>