CBCC-1280 Support for OTel tracing
Tracing updated to allow for an external tracer that creates and
finishes spans, adds tags, destroys spans.  Also, spans now have the
notion of being the 'outer' span.  That is the span that has all the
outer span tags, and is the one whose duration is used to determine
whether or not the operation has exceeded the threshold if the
ThresholdLoggingTracer is used.

If you pass in a parent that is an outer span, you must call
lcbtrace_span_finish yourself. When an operation begins, if there is no
parent span passed in, or if there is but it isn't an 'outer' span, then
the operation creates the outer span itself.

Added an example that performs a few operations which pass in various
parent spans (or none).

Still need tests.

Change-Id: Ibec5c85838e23ed21786ecc311105d98ea42316c
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/154703
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Build Bot <build@couchbase.com>