CCBC-1494: Disconnect logger after lcb_destroy invoked.
Motivation
----------
With some IO backends, the settings object and associated logger
object will be used after lcb_destroy is invoked, but this is
contrary to the ownership semantics of the logger.

Changes
-------
Force lcb_destroy to disconnect the logger before returning,
such that it is guaranteed not to be invoked after lcb_destroy
returns.

Change-Id: Ic2c940df078dcf72f4a1f0276c87ade6087c226a
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/161539
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>