tests: use vfork to avoid issues with libuv hooks
It turned out that libuv register pthread_atfork() hooks that don't play
nicely on MacOS when we use fork to run the mock process. Because we almost
immediately use execvp(), it does not make sense to deal with libuv
hooks at all, but we cannot reset or ignore them. We found that the
handlers only executed when the application uses fork(), so our options
might be vfork() and posix_spawn. This patch uses vfork(), because it
will involve minimal code change of the current code base.

Change-Id: I307adc088ba809b0f6fb1851811102e6581ca8c9
Reviewed-on: http://review.couchbase.org/c/libcouchbase/+/162549
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jake Rawsthorne <jake.rawsthorne@couchbase.com>