CCBC-1521: fix bootstrap process when client cert is used
We always pipeline error map request with HELLO request, and usually
await for hello+error_map responses, because after that goes SASL
authentication (and then optional selection of the bucket) which cannot
be completely pipelined. But in case of client certificate, we might
terminate bootstrap process too early if the bootstrap process does not
require immediate selection of the bucket.

Change-Id: Ia769a866e1f931c5758ac2e830ff062521345ca7
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/166078
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>