CCBC-1546: Add arg to allow threads to work from different rand numbers
Currently each thread in pillowfight will use the same set of random
numbers to work out which document to perform the next operation on. As
for each  thread we create we reset rand()'s seed by calling srand().

This is undesirable as for heavy read workloads this is likely to ensure
that a document is resident for most of the threads once one of them had
read a key. Thus, it does not demonstrate a true random workload.

So this patch adds an argument '--rand-space-per-thread' to allow us to
ensure that each thread has it's own random set of number which it can
use to select keys. By default this is set to false, so that by default
the behaviour of cbc-pillowfight before and after this patch is the
same.

Change-Id: I028e789562dce7b850657a5a01638e12273508ce
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/174190
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>