CCBC-1559: cbc-n1qlback: give time to IO loop in case of failure
In schenario where all query nodes suddenly failed over and/or removed
from the cluster, all requests in cbc-n1qlback will start failing.
Because the libcouchbase knows that the latest config does not have any
query nodes, it rejects all queries immediately. The single-threaded
nature of libcouchbase does not allow the IO loop to run in background,
and the code in the tool does not run `lcb_wait` in case of failure.

As a fix, we run `lcb_tick_nowait` in case of failure, which is enough
for libcouchbase to give IO loop a chance to check for any pending
events and invoke corresponding callbacks.

Change-Id: I9b426085cf7ce221ac9af7afbedae7a565dfdeaf
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/178816
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>