CCBC-1611: handle 0x0d CONFIG_ONLY status code
Treat the CONFIG_ONLY status as a signal to refresh configuration.

The new or failed over nodes are set into config-only mode, where all
data operatios will be failed with code 0x0d. It is possible that the
SDK might be using stale configuration and send requests to the node,
that is not part of the cluster anymore, so to work around this, the
library will update the configuration and retry the operation.

Change-Id: I32eab64cd99e04681a9111bc66e1e597663c8de7
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/197126
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Build Bot <build@couchbase.com>