CCBC-1616: apply wait_for_config check for all pipelines
When the connection is slow, or operation timeout is configured to
pretty low value, the SDK might not leave the event loop in case, when
lcb_wait() function is being used to execute IO operations. In this
case, regardless that actual data operation has been completed, and user
callback invoked, the library might have configuration update operation
running, and will wait for its completion, even though it is not
required to return control from lcb_wait() operation by default.

Change-Id: Ia58cb6632f3b4d17ac01b75e71b6baaddbbde2f6
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/198057
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>