CCBC-1647: Handle server-side query timeout (status 1080)
Motivation
==========
The 1080 (timeout) query status is not being handled. This means that if the server times out before the client, we return LCB_SUCCESS with an incomplete set of rows instead of LCB_ERR_TIMEOUT. Typically the client will time out before the server, however, it's possible for the server to time out first if the query grace period is set to something high enough.

Changes
=======
Return LCB_ERR_TIMEOUT when query returns the 1080 status code in the response body

Results
=======
Test passes

Change-Id: I309198fb864c0eb8e3050ac64391b30833c85a5e
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/215620
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Build Bot <build@couchbase.com>