CCBC-1651: ensure that purged packets get their memory released
mcreq_pipeline_timeout() function assumes that the packets have to be flushed to network buffers, but in case of network being configured to drop packets, the libcouchbase does not invoke flush routines as it still trying to connect sockets, so the requests would timeout without ever having chance to be written to network. In result, all buffers that associated with such requests will not be deallocated until the lcb_destroy() will be invoked on the instance.

Change-Id: Iff19b5d84bd48a14a8bb8ae70e4ba311b21b79bf
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/218172
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>