CCBC-1672: Fix protocol violation
Protocol violation by sending garbage in front of the operations for
new connections. In most of the cases, this bytes getting rejected
by the server with connection termination, and libcouchbase
successfully reconnects the pipeline after that rendering the issue
invisible for the user. But in rare cases the "garbage" might be
actual packet header, and the server does not detect problem as
everything is passing protocol checks.

To fix the issue, the mcreq_wipe_packet have to check pending list of
the pipeline and remove everything related to the current detached
packet.

Change-Id: Ib20a1c11057cf9a1ea66bbb349bcda9976f35c0f
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/232491
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>