CCBC-1672: handle case when packet gets discarded before IO calls completion handler
When the operation is getting retried, its buffers are getting relocated
to new pipeline, and removed from the old one. But if the write
operation still in the kernel and completion-based IO (e.g. windows
IOCP) is being used, the state of the buffer manager of the old pipeline
have to be adjusted.

Change-Id: I950c519546d7d29a4ce05d02eb57c14e06b2e0e2
Reviewed-on: https://review.couchbase.org/c/libcouchbase/+/233237
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>